Refactor tests from unittest to pytest
