+++ README.md
@@ -1,7 +1,8 @@
-# DriverKit
+[![codecov](https://codecov.io/gh/trpubz/MTBLDriverKit/graph/badge.svg?token=LJBMG3P28W)](https://codecov.io/gh/trpubz/MTBLDriverKit)
+# MTBL DriverKit
 
 ## Installation
-This was packaged with poetry, to install with `pip install` ensure that `DriverKit @ git+https://github.com/trpubz/SeleniumDriverKit.git@v0.2.0` syntax is located in requirements.txt
+This was packaged with poetry, to install with `pip install` ensure that `DriverKit @ git+https://github.com/trpubz/MTBLDriverKit.git@v0.3.0` syntax is located in requirements.txt
 
 ## Dependencies
-`selenium>='4.11'` is required
+`selenium>='4.16'` is required
